\name{Asia-Pacific edition data}
\alias{Batteries}
\alias{Bike}
\alias{BMI}
\alias{Carpark}
\alias{Cereals}
\alias{Curiosity}
\alias{EggCrush}
\alias{Falcon}
\alias{Fashion}
\alias{Fish}
\alias{GoGoGo}
\alias{HoldingBreath}
\alias{HumTemp}
\alias{Lifts}
\alias{Mobiles}
\alias{Movies}
\alias{PaperPlanes}
\alias{PopSong}
\alias{Reflexes}
\alias{RollingFriction}
\alias{Textbooks}
\alias{TimePerception}
\alias{VarianceOfStrength}
\alias{VisualAcuity}
\alias{WatchOut}
\docType{data}
\title{
Data from the Asia-Pacific 1st edition published November 2010}
\description{
Data sets provided by the publisher for preparation of the support documentation for using R in conjunction with the first edition of the textbook. All data sets appear in the second edition also.}
\usage{data(Mobiles)}
\format{
varies
}
\source{
Data imported from csv files provided by the publisher. Column headings were altered to ensure consistent use of capital letters, spaces, underscores, and full stops.}
\references{
Utts, J.M. and Heckard, R.F. (2010) \emph{Mind on Statistics}, Australia.
}
\examples{
data(Mobiles)
str(Mobiles)
}
\keyword{datasets}
