\name{Asia-Pacific 2nd edition data}
\alias{Aspirin}
\alias{AspirinCh11}
\alias{Blinking} 
\alias{BodyStats} 
\alias{Clasping} 
\alias{Endurance} 
\alias{Flyers} 
\alias{Holden}
\alias{LiftStairs}
\alias{Maintenance} 
\alias{OpticalIllusions} 
\alias{Pizza} 
\alias{Reading}
\alias{ServiceStation}
\alias{Study} 
\alias{Sunnies} 
\alias{TextbooksCh8}
\alias{TVAds} 
\alias{WorldRecords}

\docType{data}
\title{
Data from the Asia-Pacific 2nd edition to be published August 2013}
\description{
Data sets provided by the publisher for preparation of the support documentation for using R in conjunction with the second edition of the textbook.}
\usage{data(Aspirin)}
\format{
varies
}
\source{
Data imported from csv files provided by publisher.}
\references{
Utts, J.M. and Heckard, R.F. (2010) \emph{Mind on Statistics}, Australia.
}
\examples{
data(Aspirin)
str(Aspirin)
}
\keyword{datasets}
