\name{MindOnStats-package}
\alias{MindOnStats-package}
\alias{MindOnStats}
\docType{package}
\title{
Data for the U.S. and Asia-Pacific editions of Mind on Statistics by Utts and Heckard
}
\description{
Data were obtained from the CD-ROM accompanying the U.S. (2004) edition and from the publisher of the Asia-Pacific edition prior to its publication in 2010 for the purposes of creating support documentation for using R in conjunction with the text.}
\details{
\tabular{ll}{
Package: \tab MindOnStats\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2012-01-05\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
Data are presented as data.frames with shorter but meaningful names. Spaces in the original filenames have been removed and the initial letters of each word have been capitalised where necessary. See the index for  a complete listing of all data sets.
}
\author{
A. Jonathan R. Godfrey 
Maintainer: Jonathan Godfrey <a.j.godfrey@massey.ac.nz>
}
\references{
Utts, J.M. and Heckard, R.F. (2004) \emph{Mind on Statistics}, Thomson, Belmont, California.


}
\keyword{ package }
