\name{United States edition data}
\alias{Beans}
\alias{BearsFemale}
\alias{Cash5}
\alias{CEOData}
\alias{Ch4Heights}
\alias{Cholest}
\alias{ChugTime}
\alias{Deprived}
\alias{Fantasy5}
\alias{Fantasy5.99}
\alias{GSS93}
\alias{GSS93Samp}
\alias{HandHeight}
\alias{HappyFace}
\alias{HeightFoot}
\alias{IdealWt}
\alias{IdealWtMen}
\alias{IdealWtWomen}
\alias{Letters}
\alias{MusicCDs}
\alias{PennState1}
\alias{PennState1F}
\alias{PennState1M}
\alias{PennState2}
\alias{PennState3}
\alias{PennState4}
\alias{PennState5}
\alias{PerHouse}
\alias{Physical}
\alias{ProfBooks}
\alias{PulseMarch}
\alias{Rainfall}
\alias{SATS98}
\alias{SignDist}
\alias{SleepStudy}
\alias{SpeedLimit}
\alias{Temperature}
\alias{TurkeyPop}
\alias{UCDavis1}
\alias{UCDavis1F}
\alias{UCDavis1M}
\alias{UCDavis2}
\alias{UCDch5}
\alias{UCDWomHt}
\alias{WtHeightM}
\alias{YouthRisk}
\docType{data}
\title{
Data sets from the U.S. edition of Mind on Statistics by Utts and Heckard.
}
\description{
Data obtained from the CD-ROM accompanying the second edition, released 2004.
}
\usage{data(PennState1)}
\format{
varies
}
\source{
Data imported from tab delimited files on the CD-ROM, and files created by converting these files to csv format to handle missing values.}
\references{
Utts, J.M. and Heckard, R.F. (2004) \emph{Mind on Statistics}, Thomson, Belmont, California.
}
\examples{
data(PennState1)
str(PennState1)
}
\keyword{datasets}
